fix(ProgressBin): smoother animation
prior, we would fetch the MPRIS position every 250ms and make the animation take half that in duration. However I noticed that the progresbar felt 'choppy'. Instead of doing a smooth 0 to 100, it would do small 1% steps. After some debugging, I realized that the position would only change every 1 second so our timing was way off. It now fetches the position every 1 second. The progress animation also takes 1 second and to sync them, every time the position updates, if the animation is still playing, the starting position will be the current value instead of the previous position, which creates a smooth loading effect.