feat: MBID requirement option
one issue I have with scrobblers is that if you have mistagged songs, even slightly like missing a period or an accent, they won't be able to match the song to their database. Plus if you listen on services like youtube, it might auto play a video and it wll be logged to your scrobbling history. This option only scrobbles if it finds it in the MBID database (and fixes the metadata).

The only problem is that MB might not have the songs you listen to in their library, so it's disabled by default. It was enabled by default until I noticed that Cake Pop 2 wasn't being scrobbled.