feat(Window/CLI): emit playing on metadata update (#21)
fix: 20 (1/2)

Some apps that are gapless avoid emitting the playing property between changes (because the playing status technically didn't change) which we depend on for starting counting scrobbles. It now updates the scrobbling status when the metadata gets updated. (Only if playing is true so we can skip some checks...)

Reviewed-on: https://codeberg.org/GeopJr/Turntable/pulls/21