fix(cover): turntable fitting (#34)
fix: #33

Hoping that my maths are not wrong. Basically, I lazily appended the texture at a lower scale for turntable mode, which led to bad fittings on rectangular ratios. It now properly calculates the new values, at an as optimized branch as I could get it to. Maybe we can do it a bit more optimized since the cover is static and we can skip re-calculations on snapshotting.

Reviewed-on: https://codeberg.org/GeopJr/Turntable/pulls/34