feat: allow resizing (#65)
This was a quite a lot more work than what you'd believe. Things like the tonearm expected a static position that changed based on the cover size. Lot's of it was re-written to be more dynamic.

At the same time I did use the tonearm circle to hide the not-so-aligned arm but since I revisited the code with a clear mind, the maths were re-done and it now rotates from the correct origin point. That also allowed me to show the tonearm from the other side which makes it more realistic.

The turntable cover style wasn't calculated properly after all the attempts at fixing those fitting bugs. I now made it calculate the record vs art size dynamically which looks much better.

Some other styles and widgets had to change to better fit a more 'freeform' window size, like for example the fade/shadow style. The progress background would have an offset so it's started at the start of the fading but now that doesn't look that good when it's not at the edge so the offset was removed. It also used the window's border radius for styling but now it had to be backed-in since it's not always at the edge.

fix: #37